ResultPage has been created, the scanner for QR and other types of codes is now functional, Note: Now I have an error if someone cancel the scanner
