--- lib/ui/pages/camera_dialog.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CameraDialog extends StatelessWidget {
-  const CameraDialog({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(child: Text('Enable Gps Message'));
-  }
-}
