+++ lib/ui/pages/pages.dart
@@ -1,0 +1,2 @@
+export 'package:qbar_app/ui/pages/home/home_page.dart';
+export 'package:qbar_app/ui/pages/result/result_page.dart';
